使用 BigDecimal 代替 Double

+++ java/src/main/java/toolgood/algorithm2/AlgorithmEngine.java
@@ -249,7 +249,7 @@ public class AlgorithmEngine {
                     LastError = obj.ErrorMsg();
                     return defvalue;
                 }
-                return obj.NumberValue();
+                return obj.NumberValue().doubleValue();
             }
         } catch (final Exception ex) {
             LastError = ex.getMessage();
