+++ java/src/main/java/toolgood/flowVision/Flows/Interfaces/IInputFormulaNodeWork.java
@@ -4,6 +4,7 @@ import toolgood.flowVision.Engines.FlowEngine;
 import toolgood.flowVision.Flows.ProjectWork;
 import toolgood.flowVision.Flows.SettingFormulaItemWork;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface IInputFormulaNodeWork {
@@ -15,7 +16,7 @@ public interface IInputFormulaNodeWork {
 
     boolean IsSubsidiaryCount();
 
-    double EvaluateInputNum(FlowEngine engine) throws Exception;
+    BigDecimal EvaluateInputNum(FlowEngine engine) throws Exception;
 
     String GetMatchFormula(FlowEngine engine) throws Exception;
